+++ packages/blockchain-link/src/workers/blockfrost/types.ts
@@ -1,0 +1,6 @@
+import type { ContextType } from '../baseWorker';
+import type { BlockfrostAPI } from './websocket';
+
+export type Context = ContextType<BlockfrostAPI>;
+
+export type Request<T> = T & Context;
